+++ sample-new-architecture/src/App.tsx
@@ -19,6 +19,7 @@ import { store } from './reduxApp';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import GesturesTracingScreen from './Screens/GesturesTracingScreen';
 import { StyleSheet } from 'react-native';
+import { HttpClient } from '@sentry/integrations';
 
 const reactNavigationInstrumentation =
   new Sentry.ReactNavigationInstrumentation({
@@ -54,6 +55,16 @@ Sentry.init({
           return context;
         },
       }),
+      new HttpClient({
+        // These options are effective only in JS.
+        // This array can contain tuples of `[begin, end]` (both inclusive),
+        // Single status codes, or a combinations of both.
+        // default: [[500, 599]]
+        failedRequestStatusCodes: [[400, 599]],
+        // This array can contain Regexes or strings, or combinations of both.
+        // default: [/.*/]
+        failedRequestTargets: [/.*/],
+      }),
     );
     return integrations.filter(i => i.name !== 'Dedupe');
   },
@@ -67,6 +78,8 @@ Sentry.init({
   attachScreenshot: true,
   // Attach view hierarchy to events.
   attachViewHierarchy: true,
+  // Enables capture failed requests in JS and native.
+  enableCaptureFailedRequests: true,
   // Sets the `release` and `dist` on Sentry events. Make sure this matches EXACTLY with the values on your sourcemaps
   // otherwise they will not work.
   // release: 'myapp@1.2.3+1',